+++ models/ggml-vocab-carbon.gguf.inp
@@ -1,0 +1,25 @@
+
+__ggml_vocab_test__
+hello
+__ggml_vocab_test__
+Hello world!
+__ggml_vocab_test__
+<dna>ATCGAT</dna>
+__ggml_vocab_test__
+<dna>ATCGATCGTACG</dna>
+__ggml_vocab_test__
+<dna>atcgat</dna>
+__ggml_vocab_test__
+<dna>ATCGAN</dna>
+__ggml_vocab_test__
+<dna>ATCGA</dna>
+__ggml_vocab_test__
+<dna>ATCGATCG</dna>
+__ggml_vocab_test__
+hi <dna>ATCGAT</dna>!
+__ggml_vocab_test__
+<dna></dna>
+__ggml_vocab_test__
+<dna>ATCGAT
+__ggml_vocab_test__
+<dna>ATCGAT</dna> <dna>CGTACG</dna>
