+++ models/ggml-vocab-carbon.gguf.out
@@ -1,0 +1,13 @@
+
+14990
+9707 1879 0
+151669 152105 151670
+151669 152105 154563 151670
+151669 152105 151670
+151669 151671 151670
+151669 152104 151670
+151669 152105 154488 151670
+6023 220 151669 152105 151670 0
+151669 151670
+151669 152105
+151669 152105 151670 220 151669 154563 151670 198
